redid background image scrolling: three images all using delta to draw, but only draw two of them at a time so the third isn't rendered as it moves from the top to the bottom
zombies and trees draw with delta now, jittering seems to be fixed

+++ src/sleigher/zombie/Tree.java
@@ -6,8 +6,8 @@ import java.awt.image.BufferedImage;
 
 public class Tree {
 	
-	double x, y;
-	double lasty; //no lastx because no horizontal movement
+	float x, y;
+	float lasty; //no lastx because no horizontal movement
 	int width, height;
 	float collisionDamage;
